refactor: Public board component to use one single component
