+++ components/board-page.tsx
@@ -1,0 +1,1265 @@
+"use client";
+
+import { useState, useEffect, useRef, useMemo, useCallback } from "react";
+import { useRouter, useSearchParams } from "next/navigation";
+import { Button } from "@/components/ui/button";
+import { Checkbox } from "@/components/ui/checkbox";
+import { Input } from "@/components/ui/input";
+import { Switch } from "@/components/ui/switch";
+import { ChevronDown, Search, Copy, Trash2, X, ChevronUp, EllipsisVertical } from "lucide-react";
+import Link from "next/link";
+import { BetaBadge } from "@/components/ui/beta-badge";
+import { FilterPopover } from "@/components/ui/filter-popover";
+import { Note as NoteCard } from "@/components/note";
+
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from "@/components/ui/alert-dialog";
+import type { Note, Board, User } from "@/components/note";
+import { useTheme } from "next-themes";
+import { ProfileDropdown } from "@/components/profile-dropdown";
+import { toast } from "sonner";
+import { useUser } from "@/app/contexts/UserContext";
+import {
+  getResponsiveConfig,
+  getUniqueAuthors,
+  calculateGridLayout,
+  calculateMobileLayout,
+  filterAndSortNotes,
+} from "@/lib/utils";
+import { BoardPageSkeleton } from "@/components/board-skeleton";
+
+interface BoardPageProps {
+  params: Promise<{ id: string }>;
+  isPublic?: boolean;
+}
+
+export default function BoardPage({ params, isPublic = false }: BoardPageProps) {
+  const [board, setBoard] = useState<Board | null>(null);
+  const [notes, setNotes] = useState<Note[]>([]);
+  const { resolvedTheme } = useTheme();
+  const [allBoards, setAllBoards] = useState<Board[]>([]);
+  const [notesloading, setNotesLoading] = useState(true);
+  const { user, loading: userLoading } = useUser();
+
+  const [showBoardDropdown, setShowBoardDropdown] = useState(false);
+  const [showAddBoard, setShowAddBoard] = useState(false);
+  const [newBoardName, setNewBoardName] = useState("");
+  const [newBoardDescription, setNewBoardDescription] = useState("");
+  const [addingChecklistItem, setAddingChecklistItem] = useState<string | null>(null);
+  const [boardSettingsDialog, setBoardSettingsDialog] = useState(false);
+  const [boardSettings, setBoardSettings] = useState({
+    name: "",
+    description: "",
+    isPublic: false,
+    sendSlackUpdates: true,
+  });
+  const [copiedPublicUrl, setCopiedPublicUrl] = useState(false);
+  const [deleteConfirmDialog, setDeleteConfirmDialog] = useState(false);
+
+  const [boardId, setBoardId] = useState<string | null>(null);
+  const [isMobile, setIsMobile] = useState(false);
+  const [searchTerm, setSearchTerm] = useState("");
+  const [debouncedSearchTerm, setDebouncedSearchTerm] = useState("");
+  const [dateRange, setDateRange] = useState<{
+    startDate: Date | null;
+    endDate: Date | null;
+  }>({
+    startDate: null,
+    endDate: null,
+  });
+  const [selectedAuthor, setSelectedAuthor] = useState<string | null>(null);
+  const [errorDialog, setErrorDialog] = useState<{
+    open: boolean;
+    title: string;
+    description: string;
+  }>({ open: false, title: "", description: "" });
+
+  const pendingDeleteTimeoutsRef = useRef<Record<string, ReturnType<typeof setTimeout>>>({});
+  const boardRef = useRef<HTMLDivElement>(null);
+  const router = useRouter();
+  const searchParams = useSearchParams();
+
+  useEffect(() => {
+    if (!isPublic && !userLoading && !user) {
+      router.push("/auth/signin");
+    }
+  }, [user, userLoading, router, isPublic]);
+
+  const updateURL = useCallback(
+    (
+      newSearchTerm?: string,
+      newDateRange?: { startDate: Date | null; endDate: Date | null },
+      newAuthor?: string | null
+    ) => {
+      const params = new URLSearchParams();
+
+      const currentSearchTerm = newSearchTerm !== undefined ? newSearchTerm : searchTerm;
+      const currentDateRange = newDateRange !== undefined ? newDateRange : dateRange;
+      const currentAuthor = newAuthor !== undefined ? newAuthor : selectedAuthor;
+
+      if (currentSearchTerm) {
+        params.set("search", currentSearchTerm);
+      }
+
+      if (currentDateRange.startDate) {
+        params.set("startDate", currentDateRange.startDate.toISOString().split("T")[0]);
+      }
+
+      if (currentDateRange.endDate) {
+        params.set("endDate", currentDateRange.endDate.toISOString().split("T")[0]);
+      }
+
+      if (currentAuthor) {
+        params.set("author", currentAuthor);
+      }
+
+      const queryString = params.toString();
+      const newURL = queryString ? `?${queryString}` : window.location.pathname;
+      router.replace(newURL, { scroll: false });
+    },
+    [searchTerm, dateRange, selectedAuthor, router]
+  );
+
+  const initializeFiltersFromURL = () => {
+    if (isPublic) return;
+
+    const urlSearchTerm = searchParams.get("search") || "";
+    const urlStartDate = searchParams.get("startDate");
+    const urlEndDate = searchParams.get("endDate");
+    const urlAuthor = searchParams.get("author");
+
+    setSearchTerm(urlSearchTerm);
+
+    let startDate: Date | null = null;
+    let endDate: Date | null = null;
+
+    if (urlStartDate) {
+      const parsedStartDate = new Date(urlStartDate);
+      if (!isNaN(parsedStartDate.getTime())) {
+        startDate = parsedStartDate;
+      }
+    }
+
+    if (urlEndDate) {
+      const parsedEndDate = new Date(urlEndDate);
+      if (!isNaN(parsedEndDate.getTime())) {
+        endDate = parsedEndDate;
+      }
+    }
+
+    setDateRange({ startDate, endDate });
+    setSelectedAuthor(urlAuthor);
+  };
+
+  useEffect(() => {
+    const initializeParams = async () => {
+      const resolvedParams = await params;
+      setBoardId(resolvedParams.id);
+    };
+    initializeParams();
+  }, [params]);
+
+  useEffect(() => {
+    if (!isPublic) {
+      initializeFiltersFromURL();
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [isPublic]);
+
+  useEffect(() => {
+    if (boardId) {
+      fetchBoardData();
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [boardId]);
+
+  // Dropdown management - only for private pages
+  useEffect(() => {
+    if (isPublic) return;
+
+    const handleClickOutside = (event: MouseEvent) => {
+      if (showBoardDropdown || showAddBoard) {
+        const target = event.target as Element;
+        if (
+          !target.closest(".board-dropdown") &&
+          !target.closest(".user-dropdown") &&
+          !target.closest(".add-board-modal")
+        ) {
+          setShowBoardDropdown(false);
+          setShowAddBoard(false);
+        }
+      }
+    };
+
+    const handleKeyDown = (event: KeyboardEvent) => {
+      if (event.key === "Escape") {
+        if (addingChecklistItem) {
+          setAddingChecklistItem(null);
+        }
+        if (showBoardDropdown) {
+          setShowBoardDropdown(false);
+        }
+        if (showAddBoard) {
+          setShowAddBoard(false);
+          setNewBoardName("");
+          setNewBoardDescription("");
+        }
+      }
+    };
+
+    document.addEventListener("mousedown", handleClickOutside);
+    document.addEventListener("keydown", handleKeyDown);
+    return () => {
+      document.removeEventListener("mousedown", handleClickOutside);
+      document.removeEventListener("keydown", handleKeyDown);
+    };
+  }, [showBoardDropdown, showAddBoard, addingChecklistItem, isPublic]);
+
+  useEffect(() => {
+    let resizeTimeout: NodeJS.Timeout;
+
+    const checkResponsive = () => {
+      if (typeof window !== "undefined") {
+        const width = window.innerWidth;
+        setIsMobile(width < 768);
+
+        clearTimeout(resizeTimeout);
+        resizeTimeout = setTimeout(() => {
+          setNotes((prevNotes) => [...prevNotes]);
+        }, 50);
+      }
+    };
+
+    checkResponsive();
+    window.addEventListener("resize", checkResponsive);
+    return () => {
+      window.removeEventListener("resize", checkResponsive);
+      clearTimeout(resizeTimeout);
+    };
+  }, []);
+
+  useEffect(() => {
+    if (isPublic) {
+      setDebouncedSearchTerm(searchTerm);
+      return;
+    }
+
+    const timer = setTimeout(() => {
+      setDebouncedSearchTerm(searchTerm);
+      updateURL(searchTerm);
+    }, 1000);
+
+    return () => clearTimeout(timer);
+  }, [searchTerm, isPublic, updateURL]);
+
+  const uniqueAuthors = useMemo(() => getUniqueAuthors(notes), [notes]);
+
+  const filteredNotes = useMemo(
+    () => filterAndSortNotes(notes, debouncedSearchTerm, dateRange, selectedAuthor, isPublic ? null : user),
+    [notes, debouncedSearchTerm, dateRange, selectedAuthor, user, isPublic]
+  );
+
+  const layoutNotes = useMemo(
+    () =>
+      isMobile
+        ? calculateMobileLayout(filteredNotes, isPublic ? null : addingChecklistItem)
+        : calculateGridLayout(filteredNotes, isPublic ? null : addingChecklistItem),
+    [isMobile, filteredNotes, addingChecklistItem, isPublic]
+  );
+
+  const boardHeight = useMemo(() => {
+    if (layoutNotes.length === 0) {
+      return "calc(100vh - 64px)";
+    }
+    const maxBottom = Math.max(...layoutNotes.map((note) => note.y + note.height));
+    const minHeight = typeof window !== "undefined" && window.innerWidth < 768 ? 500 : 600;
+    const calculatedHeight = Math.max(minHeight, maxBottom + 100);
+    return `${calculatedHeight}px`;
+  }, [layoutNotes]);
+
+  const fetchBoardData = async () => {
+    try {
+      if (isPublic) {
+        const boardResponse = await fetch(`/api/boards/${boardId}`);
+
+        if (boardResponse.status === 404 || boardResponse.status === 403) {
+          setBoard(null);
+          setNotesLoading(false);
+          return;
+        }
+
+        if (boardResponse.status === 401) {
+          router.push("/auth/signin");
+          return;
+        }
+
+        if (boardResponse.ok) {
+          const { board } = await boardResponse.json();
+          if (board.isPublic) {
+            setBoard(board);
+
+            const notesResponse = await fetch(`/api/boards/${boardId}/notes`);
+            if (notesResponse.ok) {
+              const { notes } = await notesResponse.json();
+              setNotes(notes);
+            }
+          } else {
+            setBoard(null);
+          }
+        }
+      } else {
+        let allBoardsResponse: Response;
+        let notesResponse: Response | undefined;
+        let boardResponse: Response | undefined;
+
+        if (boardId === "all-notes") {
+          [allBoardsResponse, notesResponse] = await Promise.all([
+            fetch("/api/boards"),
+            fetch(`/api/boards/all-notes/notes`),
+          ]);
+
+          setBoard({
+            id: "all-notes",
+            name: "All notes",
+            description: "Notes from all boards",
+          });
+        } else if (boardId === "archive") {
+          [allBoardsResponse, notesResponse] = await Promise.all([
+            fetch("/api/boards"),
+            fetch(`/api/boards/archive/notes`),
+          ]);
+
+          setBoard({
+            id: "archive",
+            name: "Archive",
+            description: "Archived notes from all boards",
+          });
+        } else {
+          [allBoardsResponse, boardResponse, notesResponse] = await Promise.all([
+            fetch("/api/boards"),
+            fetch(`/api/boards/${boardId}`),
+            fetch(`/api/boards/${boardId}/notes`),
+          ]);
+        }
+
+        if (allBoardsResponse.ok) {
+          const { boards } = await allBoardsResponse.json();
+          setAllBoards(boards);
+        }
+
+        if (boardResponse && boardResponse.ok) {
+          const { board } = await boardResponse.json();
+          setBoard(board);
+          setBoardSettings({
+            name: board.name,
+            description: board.description || "",
+            isPublic: (board as { isPublic?: boolean })?.isPublic ?? false,
+            sendSlackUpdates: (board as { sendSlackUpdates?: boolean })?.sendSlackUpdates ?? true,
+          });
+        }
+
+        if (notesResponse && notesResponse.ok) {
+          const { notes } = await notesResponse.json();
+          setNotes(notes);
+        }
+
+        if (boardId && boardId !== "all-notes") {
+          try {
+            localStorage.setItem("gumboard-last-visited-board", boardId);
+          } catch (error) {
+            console.warn("Failed to save last visited board:", error);
+          }
+        }
+      }
+    } catch (error) {
+      console.error("Error fetching board data:", error);
+      setBoard(null);
+    } finally {
+      setNotesLoading(false);
+    }
+  };
+
+  const handleUpdateNoteFromComponent = async (updatedNote: Note) => {
+    const currentNote = notes.find((n) => n.id === updatedNote.id);
+    if (!currentNote) return;
+
+    setNotes((prev) => prev.map((n) => (n.id === updatedNote.id ? updatedNote : n)));
+  };
+
+  const handleAddNote = async (targetBoardId?: string) => {
+    if (boardId === "all-notes" && !targetBoardId) {
+      setErrorDialog({
+        open: true,
+        title: "Board selection required",
+        description: "Please select a board to add the note to",
+      });
+      return;
+    }
+
+    try {
+      const actualTargetBoardId = boardId === "all-notes" ? targetBoardId : boardId;
+      const isAllNotesView = boardId === "all-notes";
+
+      const response = await fetch(
+        `/api/boards/${isAllNotesView ? "all-notes" : actualTargetBoardId}/notes`,
+        {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({
+            checklistItems: [],
+            ...(isAllNotesView && { boardId: targetBoardId }),
+          }),
+        }
+      );
+
+      if (response.ok) {
+        const { note } = await response.json();
+        setNotes((prev) => [...prev, note]);
+        setAddingChecklistItem(note.id);
+        if (searchTerm.trim() || dateRange.startDate || dateRange.endDate || selectedAuthor) {
+          setSearchTerm("");
+          setDebouncedSearchTerm("");
+          setDateRange({ startDate: null, endDate: null });
+          setSelectedAuthor(null);
+          updateURL("", { startDate: null, endDate: null }, null);
+        }
+      }
+    } catch (error) {
+      console.error("Error creating note:", error);
+    }
+  };
+
+  const handleCopyNote = async (originalNote: Note) => {
+    try {
+      const targetBoardId = boardId === "all-notes" ? originalNote.boardId : boardId;
+      const isAllNotesView = boardId === "all-notes";
+
+      const checklistItems =
+        originalNote.checklistItems?.map((item, index) => ({
+          content: item.content,
+          checked: item.checked,
+          order: index,
+        })) || [];
+
+      const response = await fetch(
+        `/api/boards/${isAllNotesView ? "all-notes" : targetBoardId}/notes`,
+        {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({
+            color: originalNote.color,
+            checklistItems,
+            ...(isAllNotesView && { boardId: targetBoardId }),
+          }),
+        }
+      );
+
+      if (response.ok) {
+        const { note } = await response.json();
+        setNotes((prev) => [...prev, note]);
+      }
+    } catch (error) {
+      console.error("Error copying note:", error);
+    }
+  };
+
+  const handleDeleteNote = (noteId: string) => {
+    const noteToDelete = notes.find((n) => n.id === noteId);
+    if (!noteToDelete) return;
+
+    const targetBoardId = noteToDelete.board?.id ?? noteToDelete.boardId;
+
+    setNotes((prev) => prev.filter((n) => n.id !== noteId));
+
+    const timeoutId = setTimeout(async () => {
+      try {
+        const response = await fetch(`/api/boards/${targetBoardId}/notes/${noteId}`, {
+          method: "DELETE",
+        });
+        if (!response.ok) {
+          setNotes((prev) => [noteToDelete, ...prev]);
+          const errorData = await response.json().catch(() => null);
+          setErrorDialog({
+            open: true,
+            title: "Failed to delete note",
+            description: errorData?.error || "Failed to delete note",
+          });
+        }
+      } catch (error) {
+        console.error("Error deleting note:", error);
+        setNotes((prev) => [noteToDelete, ...prev]);
+        setErrorDialog({
+          open: true,
+          title: "Failed to delete note",
+          description: "Failed to delete note",
+        });
+      } finally {
+        delete pendingDeleteTimeoutsRef.current[noteId];
+      }
+    }, 4000);
+    pendingDeleteTimeoutsRef.current[noteId] = timeoutId;
+
+    toast("Note deleted", {
+      action: {
+        label: "Undo",
+        onClick: () => {
+          const t = pendingDeleteTimeoutsRef.current[noteId];
+          if (t) {
+            clearTimeout(t);
+            delete pendingDeleteTimeoutsRef.current[noteId];
+          }
+          setNotes((prev) => [noteToDelete, ...prev]);
+        },
+      },
+      duration: 4000,
+    });
+  };
+
+  const handleArchiveNote = async (noteId: string) => {
+    try {
+      const currentNote = notes.find((n) => n.id === noteId);
+      if (!currentNote) return;
+
+      const targetBoardId = currentNote?.board?.id ?? currentNote.boardId;
+
+      setNotes((prev) => prev.filter((n) => n.id !== noteId));
+
+      const response = await fetch(`/api/boards/${targetBoardId}/notes/${noteId}`, {
+        method: "PUT",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ archivedAt: new Date().toISOString() }),
+      });
+
+      if (!response.ok) {
+        setNotes((prev) => [...prev, currentNote]);
+        setErrorDialog({
+          open: true,
+          title: "Archive Failed",
+          description: "Failed to archive note. Please try again.",
+        });
+      }
+    } catch (error) {
+      console.error("Error archiving note:", error);
+    }
+  };
+
+  const handleUnarchiveNote = async (noteId: string) => {
+    try {
+      const currentNote = notes.find((n) => n.id === noteId);
+      if (!currentNote) return;
+
+      const targetBoardId = currentNote.board?.id ?? currentNote.boardId;
+      if (!targetBoardId) return;
+
+      setNotes((prev) => prev.filter((n) => n.id !== noteId));
+
+      const response = await fetch(`/api/boards/${targetBoardId}/notes/${noteId}`, {
+        method: "PUT",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ archivedAt: null }),
+      });
+
+      if (!response.ok) {
+        setNotes((prev) => [...prev, currentNote]);
+        setErrorDialog({
+          open: true,
+          title: "Unarchive Failed",
+          description: "Failed to unarchive note. Please try again.",
+        });
+      }
+    } catch (error) {
+      console.error("Error unarchiving note:", error);
+    }
+  };
+
+  const handleAddBoard = async (e: React.FormEvent) => {
+    e.preventDefault();
+    if (!newBoardName.trim()) return;
+
+    try {
+      const response = await fetch("/api/boards", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          name: newBoardName,
+          description: newBoardDescription,
+        }),
+      });
+
+      if (response.ok) {
+        const { board } = await response.json();
+        setAllBoards([board, ...allBoards]);
+        setNewBoardName("");
+        setNewBoardDescription("");
+        setShowAddBoard(false);
+        setShowBoardDropdown(false);
+        router.push(`/boards/${board.id}`);
+      } else {
+        const errorData = await response.json();
+        setErrorDialog({
+          open: true,
+          title: "Failed to create board",
+          description: errorData.error || "Failed to create board",
+        });
+      }
+    } catch (error) {
+      console.error("Error creating board:", error);
+      setErrorDialog({
+        open: true,
+        title: "Failed to create board",
+        description: "Failed to create board",
+      });
+    }
+  };
+
+  const handleUpdateBoardSettings = async (settings: {
+    name?: string;
+    description?: string;
+    isPublic?: boolean;
+    sendSlackUpdates: boolean;
+  }) => {
+    try {
+      const response = await fetch(`/api/boards/${boardId}`, {
+        method: "PUT",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify(settings),
+      });
+
+      if (response.ok) {
+        const { board } = await response.json();
+        setBoard(board);
+
+        setAllBoards((prevBoards) => prevBoards.map((b) => (b.id === board.id ? board : b)));
+
+        setBoardSettings({
+          name: board.name,
+          description: board.description || "",
+          isPublic: (board as { isPublic?: boolean })?.isPublic ?? false,
+          sendSlackUpdates: (board as { sendSlackUpdates?: boolean })?.sendSlackUpdates ?? true,
+        });
+        setBoardSettingsDialog(false);
+      }
+    } catch (error) {
+      console.error("Error updating board settings:", error);
+    }
+  };
+
+  const handleCopyPublicUrl = async () => {
+    const publicUrl = `${window.location.origin}/public/boards/${boardId}`;
+    try {
+      await navigator.clipboard.writeText(publicUrl);
+      setCopiedPublicUrl(true);
+      setTimeout(() => setCopiedPublicUrl(false), 2000);
+    } catch (error) {
+      console.error("Failed to copy URL:", error);
+    }
+  };
+
+  const handleDeleteBoard = async () => {
+    try {
+      const response = await fetch(`/api/boards/${boardId}`, {
+        method: "DELETE",
+      });
+
+      if (response.ok) {
+        router.push("/dashboard");
+      } else {
+        const errorData = await response.json();
+        setErrorDialog({
+          open: true,
+          title: "Failed to delete board",
+          description: errorData.error || "Failed to delete board",
+        });
+      }
+    } catch (error) {
+      console.error("Error deleting board:", error);
+      setErrorDialog({
+        open: true,
+        title: "Failed to delete board",
+        description: "Failed to delete board",
+      });
+    }
+    setDeleteConfirmDialog(false);
+  };
+
+  if (userLoading || notesloading) {
+    return <BoardPageSkeleton />;
+  }
+
+  if (!board && boardId !== "all-notes" && boardId !== "archive") {
+    return (
+      <div className="min-h-screen bg-zinc-100 dark:bg-zinc-800 bg-dots flex items-center justify-center">
+        <div className="text-center">
+          <h1 className="text-2xl font-bold mb-2">Board not found</h1>
+          <p className="text-gray-600 dark:text-gray-400 mb-4">
+            {isPublic
+              ? "This board doesn't exist or is not publicly accessible."
+              : "This board doesn't exist or you don't have access to it."
+            }
+          </p>
+          <Button asChild>
+            <Link href={"/"}>
+              Go to Gumboard
+            </Link>
+          </Button>
+        </div>
+      </div>
+    );
+  }
+
+  return (
+    <div className="min-h-screen max-w-screen bg-zinc-100 dark:bg-zinc-800 bg-dots">
+      <div>
+        <div className="mx-4 flex flex-wrap sm:flex-nowrap justify-between items-center h-auto sm:h-16 p-2 sm:p-0">
+          <div className="bg-white dark:bg-zinc-900 shadow-sm border border-zinc-100 rounded-lg dark:border-zinc-800 mt-2 py-2 px-3 flex flex-wrap sm:flex-nowrap items-center w-full sm:w-auto">
+            <Link href={"/"} className="flex-shrink-0 pl-1">
+              <h1 className="text-xl font-bold text-zinc-900 dark:text-zinc-100 flex items-center gap-3">
+                Gumboard
+                <BetaBadge />
+              </h1>
+            </Link>
+            <div className="h-6 w-px my-1.5 mx-4 bg-zinc-100 dark:bg-zinc-700" />
+
+            {isPublic ? (
+              <>
+                <div className="flex-1 mr-0 sm:flex-none">
+                  <div className="text-sm font-semibold text-foreground dark:text-zinc-100">
+                    {board?.name}
+                  </div>
+                </div>
+                <div className="h-6 w-px m-1.5 bg-zinc-100 dark:bg-zinc-700" />
+                <span className="inline-flex items-center px-2 py-1 rounded-full text-xs font-medium bg-green-100 text-green-800 dark:bg-green-900 dark:text-green-200">
+                  Public
+                </span>
+              </>
+            ) : (
+              <div className="relative board-dropdown min-w-32 sm:max-w-64 col-span-2 sm:col-span-1">
+                <Button
+                  variant="ghost"
+                  onClick={() => setShowBoardDropdown(!showBoardDropdown)}
+                  className="flex items-center justify-between px-2 py-2 w-full"
+                >
+                  <div className="min-w-0">
+                    <div className="text-sm font-semibold text-foreground dark:text-zinc-100 truncate">
+                      {boardId === "all-notes"
+                        ? "All notes"
+                        : boardId === "archive"
+                          ? "Archive"
+                          : board?.name}
+                    </div>
+                  </div>
+                  {showBoardDropdown ? (
+                    <ChevronUp className="w-4 h-4 text-muted-foreground dark:text-zinc-400 transition-transform" />
+                  ) : (
+                    <ChevronDown className="w-4 h-4 text-muted-foreground dark:text-zinc-400 transition-transform" />
+                  )}
+                </Button>
+
+                {showBoardDropdown && (
+                  <div className="fixed sm:absolute left-0 mt-1 w-full sm:w-64 bg-white dark:bg-zinc-900 rounded-lg shadow-lg border border-zinc-100 dark:border-zinc-800 z-50 max-h-80 overflow-y-auto">
+                    <div className="p-2 flex flex-col gap-1">
+                      {allBoards.map((b) => (
+                        <Link
+                          key={b.id}
+                          href={`/boards/${b.id}`}
+                          className={`rounded-lg block font-medium px-3 py-1.5 text-sm hover:bg-zinc-100 dark:hover:bg-zinc-800 dark:hover:text-white ${b.id === boardId
+                            ? "bg-sky-50 dark:bg-sky-600 text-foreground dark:text-zinc-100 font-semibold"
+                            : "text-foreground dark:text-zinc-100"
+                            }`}
+                          onClick={() => setShowBoardDropdown(false)}
+                        >
+                          <div>{b.name}</div>
+                        </Link>
+                      ))}
+
+                      {allBoards.length > 0 && (
+                        <div className="border-t border-zinc-100 dark:border-zinc-800 my-1"></div>
+                      )}
+
+                      <Link
+                        href="/boards/all-notes"
+                        className={`rounded-lg font-medium block px-3 py-1.5 text-sm hover:bg-zinc-100 dark:hover:bg-zinc-800 ${boardId === "all-notes"
+                          ? "bg-zinc-100 dark:bg-zinc-800 dark:text-white font-semibold"
+                          : "text-foreground dark:text-white"
+                          }`}
+                        onClick={() => setShowBoardDropdown(false)}
+                      >
+                        <div>All notes</div>
+                      </Link>
+
+                      <Link
+                        href="/boards/archive"
+                        className={`rounded-lg block font-medium px-3 py-1.5 text-sm hover:bg-zinc-100 dark:hover:bg-zinc-800 ${boardId === "archive"
+                          ? "bg-zinc-100 dark:bg-zinc-800 dark:text-white font-semibold"
+                          : "text-foreground dark:text-white"
+                          }`}
+                        onClick={() => setShowBoardDropdown(false)}
+                      >
+                        <div>All archived</div>
+                      </Link>
+                      <div className="border-t border-zinc-100 dark:border-zinc-800 my-1"></div>
+                      <Button
+                        variant="outline"
+                        size="sm"
+                        onClick={() => {
+                          setShowAddBoard(true);
+                          setShowBoardDropdown(false);
+                        }}
+                        className="flex items-center w-full px-4 py-2"
+                      >
+                        <span className="font-medium">Create new board</span>
+                      </Button>
+                    </div>
+                  </div>
+                )}
+              </div>
+            )}
+
+            <div className="h-6 w-px my-1.5 mx-2 bg-zinc-100 dark:bg-zinc-700" />
+
+            <div className="flex flex-nowrap space-x-2">
+              <div className="relative board-dropdown" data-slot="filter-popover">
+                <FilterPopover
+                  startDate={dateRange.startDate}
+                  endDate={dateRange.endDate}
+                  onDateRangeChange={(startDate, endDate) => {
+                    const newDateRange = { startDate, endDate };
+                    setDateRange(newDateRange);
+                    if (!isPublic) updateURL(undefined, newDateRange);
+                  }}
+                  selectedAuthor={selectedAuthor}
+                  authors={uniqueAuthors}
+                  onAuthorChange={(authorId) => {
+                    setSelectedAuthor(authorId);
+                    if (!isPublic) updateURL(undefined, undefined, authorId);
+                  }}
+                  className={isPublic ? "w-fit" : "h-9"}
+                />
+              </div>
+
+              {!isPublic && boardId !== "all-notes" && boardId !== "archive" && (
+                <Button
+                  variant="ghost"
+                  size="sm"
+                  onClick={() => {
+                    setBoardSettings({
+                      name: board?.name || "",
+                      description: board?.description || "",
+                      isPublic: (board as { isPublic?: boolean })?.isPublic ?? false,
+                      sendSlackUpdates:
+                        (board as { sendSlackUpdates?: boolean })?.sendSlackUpdates ?? true,
+                    });
+                    setBoardSettingsDialog(true);
+                  }}
+                  aria-label="Board settings"
+                  title="Board settings"
+                  className="flex items-center size-9"
+                >
+                  <EllipsisVertical className="size-4" />
+                </Button>
+              )}
+            </div>
+          </div>
+
+          <div className="bg-white dark:bg-zinc-900 shadow-sm border border-zinc-100 rounded-lg dark:border-zinc-800 mt-2 py-2 px-3 flex flex-wrap sm:flex-nowrap items-center sm:space-x-3 w-full sm:w-auto">
+            <div className="flex justify-between gap-2 items-center w-full sm:w-auto">
+              <div className="relative flex-1 sm:flex-none min-w-[150px]">
+                <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                  <Search className="h-4 w-4 text-muted-foreground dark:text-zinc-400" />
+                </div>
+                <input
+                  aria-label="Search notes"
+                  type="text"
+                  placeholder="Search notes..."
+                  value={searchTerm}
+                  onChange={(e) => setSearchTerm(e.target.value)}
+                  className="w-full sm:w-64 pl-10 pr-8 py-2 border border-zinc-100 dark:border-zinc-800 rounded-md focus:outline-none focus:ring-2 focus:ring-sky-600 dark:focus:ring-sky-600 focus:border-transparent text-sm bg-background dark:bg-zinc-900 text-foreground dark:text-zinc-100 placeholder:text-muted-foreground dark:placeholder:text-zinc-400"
+                />
+                {searchTerm && !isPublic && (
+                  <Button
+                    onClick={() => {
+                      setSearchTerm("");
+                      setDebouncedSearchTerm("");
+                      updateURL("");
+                    }}
+                    className="absolute top-[5px] right-1 size-7 flex items-center text-muted-foreground dark:text-zinc-400 hover:text-white dark:hover:text-zinc-100 cursor-pointer bg-transparent"
+                  >
+                    <X className="h-4 w-4" />
+                  </Button>
+                )}
+              </div>
+
+              {!isPublic && (
+                <Button
+                  onClick={() => {
+                    if (boardId === "all-notes" && allBoards.length > 0) {
+                      handleAddNote(allBoards[0].id);
+                    } else {
+                      handleAddNote();
+                    }
+                  }}
+                  disabled={boardId === "archive"}
+                >
+                  <span>Add note</span>
+                </Button>
+              )}
+
+              {user ? (
+                <ProfileDropdown user={user} />
+              ) : (
+                <Link href="/auth/signin">
+                  <Button variant="outline" size="sm">
+                    Sign in
+                  </Button>
+                </Link>
+              )}
+            </div>
+          </div>
+        </div>
+      </div>
+
+      <div
+        ref={boardRef}
+        className="relative w-full"
+        style={{
+          height: boardHeight,
+          minHeight: "calc(100vh - 64px)",
+        }}
+      >
+        <div className="relative w-full h-full">
+          {layoutNotes.map((note) => (
+            <NoteCard
+              key={note.id}
+              note={note as Note}
+              currentUser={isPublic ? undefined : (user as User)}
+              readonly={isPublic}
+              onUpdate={!isPublic ? handleUpdateNoteFromComponent : undefined}
+              onDelete={!isPublic ? handleDeleteNote : undefined}
+              onArchive={!isPublic && boardId !== "archive" ? handleArchiveNote : undefined}
+              onUnarchive={!isPublic && boardId === "archive" ? handleUnarchiveNote : undefined}
+              onCopy={!isPublic ? handleCopyNote : undefined}
+              showBoardName={boardId === "all-notes" || boardId === "archive"}
+              className="shadow-md shadow-black/10"
+              style={{
+                position: "absolute",
+                left: note.x,
+                top: note.y,
+                width: note.width,
+                height: note.height,
+                padding: `${getResponsiveConfig().notePadding}px`,
+                backgroundColor: resolvedTheme === "dark" ? "#18181B" : note.color,
+              }}
+            />
+          ))}
+        </div>
+
+        {filteredNotes.length === 0 && (
+          <div className="absolute inset-0 flex items-center justify-center">
+            <div className="text-center">
+              <div className="text-6xl mb-4">📝</div>
+              <h3 className="text-lg font-medium text-gray-900 dark:text-gray-100 mb-2">
+                No notes found
+              </h3>
+              <p className="text-gray-500 dark:text-gray-400">
+                {searchTerm || selectedAuthor || dateRange.startDate || dateRange.endDate
+                  ? "Try adjusting your filters to see more notes."
+                  : "This board doesn't have any notes yet."}
+              </p>
+
+              {!isPublic && (searchTerm || dateRange.startDate || dateRange.endDate || selectedAuthor) && (
+                <Button
+                  onClick={() => {
+                    setSearchTerm("");
+                    setDebouncedSearchTerm("");
+                    setDateRange({ startDate: null, endDate: null });
+                    setSelectedAuthor(null);
+                    updateURL("", { startDate: null, endDate: null }, null);
+                  }}
+                  variant="outline"
+                  className="flex items-center space-x-2 cursor-pointer mt-4"
+                >
+                  <span>Clear All Filters</span>
+                </Button>
+              )}
+            </div>
+          </div>
+        )}
+      </div>
+
+      {!isPublic && (
+        <>
+          {showAddBoard && (
+            <div
+              className="fixed inset-0 flex items-center justify-center z-50 p-4 bg-black/40 dark:bg-black/70 backdrop-blur-sm add-board-modal"
+              onClick={() => {
+                setShowAddBoard(false);
+                setNewBoardName("");
+                setNewBoardDescription("");
+              }}
+            >
+              <div
+                className="bg-white dark:bg-zinc-950 bg-opacity-95 dark:bg-opacity-95 rounded-xl p-5 sm:p-7 w-full max-w-sm sm:max-w-md shadow-2xl border border-gray-200 dark:border-zinc-800"
+                onClick={(e) => e.stopPropagation()}
+              >
+                <h3 className="text-lg font-semibold mb-4 text-foreground dark:text-zinc-100">
+                  Create new board
+                </h3>
+                <form onSubmit={handleAddBoard}>
+                  <div className="space-y-4">
+                    <div>
+                      <label className="block text-sm font-medium text-foreground dark:text-zinc-200 mb-1">
+                        Board name
+                      </label>
+                      <Input
+                        type="text"
+                        value={newBoardName}
+                        onChange={(e) => setNewBoardName(e.target.value)}
+                        placeholder="Enter board name"
+                        required
+                        className="bg-white dark:bg-zinc-900 text-foreground dark:text-zinc-100 border border-gray-200 dark:border-zinc-700"
+                      />
+                    </div>
+                    <div>
+                      <label className="block text-sm font-medium text-foreground dark:text-zinc-200 mb-1">
+                        Description (optional)
+                      </label>
+                      <Input
+                        type="text"
+                        value={newBoardDescription}
+                        onChange={(e) => setNewBoardDescription(e.target.value)}
+                        placeholder="Enter board description"
+                        className="bg-white dark:bg-zinc-900 text-foreground dark:text-zinc-100 border border-gray-200 dark:border-zinc-700"
+                      />
+                    </div>
+                  </div>
+                  <div className="flex justify-end space-x-3 mt-6">
+                    <Button
+                      type="button"
+                      variant="outline"
+                      onClick={() => {
+                        setShowAddBoard(false);
+                        setNewBoardName("");
+                        setNewBoardDescription("");
+                      }}
+                      className="bg-white dark:bg-zinc-900 text-foreground dark:text-zinc-100 border border-gray-300 dark:border-zinc-700 hover:bg-zinc-100 hover:text-foreground hover:border-gray-300 dark:hover:bg-zinc-800"
+                    >
+                      Cancel
+                    </Button>
+                    <Button
+                      type="submit"
+                      className="bg-blue-600 hover:bg-blue-700 text-zinc-100 dark:bg-blue-500 dark:hover:bg-blue-600 dark:text-zinc-100"
+                    >
+                      Create board
+                    </Button>
+                  </div>
+                </form>
+              </div>
+            </div>
+          )}
+
+          <AlertDialog open={boardSettingsDialog} onOpenChange={setBoardSettingsDialog}>
+            <AlertDialogContent className="bg-white dark:bg-zinc-950 border border-gray-200 dark:border-zinc-800 p-4 lg:p-6">
+              <AlertDialogHeader>
+                <AlertDialogTitle className="text-foreground dark:text-zinc-100">
+                  Board settings
+                </AlertDialogTitle>
+                <AlertDialogDescription className="text-muted-foreground dark:text-zinc-400">
+                  Configure settings for "{board?.name}" board.
+                </AlertDialogDescription>
+              </AlertDialogHeader>
+
+              <div className="py-4 space-y-4">
+                <div>
+                  <label className="block text-sm font-medium text-foreground dark:text-zinc-200 mb-1">
+                    Board name
+                  </label>
+                  <Input
+                    type="text"
+                    value={boardSettings.name}
+                    onChange={(e) => setBoardSettings((prev) => ({ ...prev, name: e.target.value }))}
+                    placeholder="Enter board name"
+                    className="bg-white dark:bg-zinc-900 text-foreground dark:text-zinc-100 border border-gray-200 dark:border-zinc-700"
+                  />
+                </div>
+                <div>
+                  <label className="block text-sm font-medium text-foreground dark:text-zinc-200 mb-1">
+                    Description (optional)
+                  </label>
+                  <Input
+                    type="text"
+                    value={boardSettings.description}
+                    onChange={(e) =>
+                      setBoardSettings((prev) => ({ ...prev, description: e.target.value }))
+                    }
+                    placeholder="Enter board description"
+                    className="bg-white dark:bg-zinc-900 text-foreground dark:text-zinc-100 border border-gray-200 dark:border-zinc-700"
+                  />
+                </div>
+                <div className="space-y-4">
+                  <div className="flex items-center space-x-2">
+                    <Switch
+                      id="isPublic"
+                      checked={boardSettings.isPublic}
+                      onCheckedChange={(checked) =>
+                        setBoardSettings((prev) => ({ ...prev, isPublic: checked as boolean }))
+                      }
+                    />
+                    <label
+                      htmlFor="isPublic"
+                      className="text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70 text-foreground dark:text-zinc-100"
+                    >
+                      Make board public
+                    </label>
+                  </div>
+                  <p className="text-xs text-muted-foreground dark:text-zinc-400 mt-1 ml-6">
+                    When enabled, anyone with the link can view this board
+                  </p>
+
+                  {boardSettings.isPublic && (
+                    <div className="ml-6 p-3 bg-gray-50 dark:bg-zinc-800 rounded-md">
+                      <div className="flex items-center justify-between">
+                        <div>
+                          <p className="text-sm font-medium text-foreground dark:text-zinc-100">
+                            Public link
+                          </p>
+                          <p className="text-xs text-muted-foreground dark:text-zinc-400 break-all">
+                            {typeof window !== "undefined"
+                              ? `${window.location.origin}/public/boards/${boardId}`
+                              : ""}
+                          </p>
+                        </div>
+                        <Button
+                          onClick={handleCopyPublicUrl}
+                          size="sm"
+                          variant="outline"
+                          className="ml-3 flex items-center space-x-1"
+                        >
+                          {copiedPublicUrl ? (
+                            <>
+                              <span className="text-xs">✓</span>
+                              <span className="text-xs">Copied!</span>
+                            </>
+                          ) : (
+                            <>
+                              <Copy className="w-3 h-3" />
+                              <span className="text-xs">Copy</span>
+                            </>
+                          )}
+                        </Button>
+                      </div>
+                    </div>
+                  )}
+                </div>
+
+                <div className="flex items-center space-x-2">
+                  <Checkbox
+                    id="sendSlackUpdates"
+                    checked={boardSettings.sendSlackUpdates}
+                    onCheckedChange={(checked) =>
+                      setBoardSettings((prev) => ({ ...prev, sendSlackUpdates: checked as boolean }))
+                    }
+                    className="border-slate-500 bg-white/50 dark:bg-zinc-800 dark:text-zinc-300 dark:border-zinc-600 mt-1"
+                  />
+                  <label
+                    htmlFor="sendSlackUpdates"
+                    className="text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70 text-foreground dark:text-zinc-100"
+                  >
+                    Send updates to Slack
+                  </label>
+                </div>
+                <p className="text-xs text-muted-foreground dark:text-zinc-400 mt-1 ml-6">
+                  When enabled, note updates will be sent to your organization's Slack channel
+                </p>
+              </div>
+
+              <AlertDialogFooter className="flex !flex-row justify-between">
+                <Button
+                  onClick={() => setDeleteConfirmDialog(true)}
+                  variant="destructive"
+                  className="mr-auto bg-red-600 hover:bg-red-700 text-white dark:bg-red-600 dark:hover:bg-red-700"
+                >
+                  <Trash2 className="w-4 h-4" />
+                  Delete <span className="hidden lg:inline">Board</span>
+                </Button>
+                <div className="flex space-x-2 items-center">
+                  <AlertDialogCancel className="border-gray-400 text-foreground dark:text-zinc-100 dark:border-zinc-700 hover:bg-zinc-100 hover:text-foreground hover:border-gray-200 dark:hover:bg-zinc-800">
+                    Cancel
+                  </AlertDialogCancel>
+                  <AlertDialogAction
+                    onClick={() => handleUpdateBoardSettings(boardSettings)}
+                    className="bg-blue-600 hover:bg-blue-700 text-white dark:bg-blue-500 dark:hover:bg-blue-600 dark:text-white"
+                  >
+                    Save settings
+                  </AlertDialogAction>
+                </div>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialog>
+
+          <AlertDialog open={deleteConfirmDialog} onOpenChange={setDeleteConfirmDialog}>
+            <AlertDialogContent className="bg-white dark:bg-zinc-950 border border-gray-200 dark:border-zinc-800">
+              <AlertDialogHeader>
+                <AlertDialogTitle className="text-foreground dark:text-zinc-100">
+                  Delete Board
+                </AlertDialogTitle>
+                <AlertDialogDescription className="text-muted-foreground dark:text-zinc-400">
+                  Are you sure you want to delete "{board?.name}"? This action cannot be
+                  undone and will permanently delete all notes in this board.
+                </AlertDialogDescription>
+              </AlertDialogHeader>
+              <AlertDialogFooter>
+                <AlertDialogCancel className="bg-white dark:bg-zinc-900 text-foreground dark:text-zinc-100 border border-gray-200 dark:border-zinc-700 hover:bg-zinc-100 dark:hover:bg-zinc-800">
+                  Cancel
+                </AlertDialogCancel>
+                <AlertDialogAction
+                  onClick={handleDeleteBoard}
+                  className="bg-red-600 hover:bg-red-700 text-white dark:bg-red-600 dark:hover:bg-red-700"
+                >
+                  Delete Board
+                </AlertDialogAction>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialog>
+        </>
+      )}
+
+      <AlertDialog
+        open={errorDialog.open}
+        onOpenChange={(open) => setErrorDialog({ open, title: "", description: "" })}
+      >
+        <AlertDialogContent className="bg-white dark:bg-zinc-950 border border-gray-200 dark:border-zinc-800">
+          <AlertDialogHeader>
+            <AlertDialogTitle className="text-foreground dark:text-zinc-100">
+              {errorDialog.title}
+            </AlertDialogTitle>
+            <AlertDialogDescription className="text-muted-foreground dark:text-zinc-400">
+              {errorDialog.description}
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogAction
+              onClick={() => setErrorDialog({ open: false, title: "", description: "" })}
+              className="bg-red-600 hover:bg-red-700 text-white dark:bg-red-600 dark:hover:bg-red-700"
+            >
+              OK
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
+    </div>
+  );
+}
